Add RPM calculator
Right now the code is really ugly, and not DRY. I would like do to some
more work to it, but I think this is a good commit with the code
working.

I think the Kata online has an error:
https://codingdojo.org/kata/RPN/

The last example multiplies by 3 at the end, this happens to show the
correct answer because multiplication is commutative, however if we were
subtracting I think the answer would be wrong.

I changed the tests to match what I think is correct.