Refactor the rpm calculator
I grouped similar logic so that it wasn't repeated as much. I also made
some private methods which makes the class a bit easier to read.

There are still a lot of params being passed around. If we made a new
class that initalized some of these params it might be cleaner.